Extracts dialog widget
